+++ lifelines/fitters/coxph_fitter.py
@@ -99,7 +99,7 @@ class CoxPHFitter(RegressionFitter, ProportionalHazardMixin):
         The event_observed variable provided
     weights: Series
         The event_observed variable provided
-    variance_matrix_ : numpy array
+    variance_matrix_ : DataFrame
         The variance matrix of the coefficients
     strata: list
         the strata provided
@@ -209,7 +209,7 @@ class CoxPHFitter(RegressionFitter, ProportionalHazardMixin):
             "The Robust Inference for the Cox Proportional Hazards Model", Journal of the American Statistical Association, Vol. 84, No. 408 (Dec., 1989), pp. 1074- 1078
 
         formula: str, optional
-            an Wilkinson formula, like in R and statsmodels, for the RHS. If left as None, all columns not assigned as durations, weights, etc. are used.
+            an Wilkinson formula, like in R and statsmodels, for the right-hand-side. If left as None, all columns not assigned as durations, weights, etc. are used.
 
         batch_mode: bool, optional
             enabling batch_mode can be faster for datasets with a large number of ties. If left as None, lifelines will choose the best option.
@@ -265,7 +265,6 @@ class CoxPHFitter(RegressionFitter, ProportionalHazardMixin):
             cph = CoxPHFitter()
             cph.fit(df, 'T', 'E', strata=['month', 'age'], robust=True, weights_col='weights')
             cph.print_summary()
-            cph.predict_median(df)
 
         """
         self.strata = utils.coalesce(strata, self.strata)
@@ -358,7 +357,7 @@ class CoxPHFitter(RegressionFitter, ProportionalHazardMixin):
         style: string
             {html, ascii, latex}
         columns:
-            only display a subset of `summary` columns. Default all.
+            only display a subset of ``summary`` columns. Default all.
         kwargs:
             print additional metadata in the output (useful to provide model names, dataset names, etc.) when comparing
             multiple outputs.
@@ -530,7 +529,7 @@ class SemiParametricPHFitter(ProportionalHazardMixin, SemiParametricRegressionFi
         The event_observed variable provided
     weights: Series
         The event_observed variable provided
-    variance_matrix_ : numpy array
+    variance_matrix_ : DataFrame
         The variance matrix of the coefficients
     strata: list
         the strata provided
@@ -2115,7 +2114,7 @@ See https://stats.stackexchange.com/q/11109/11867 for more.\n",
 
     def plot_covariate_groups(*args, **kwargs):
         """
-        Deprecated as of v0.25.0. Use plot_partial_effects_on_outcome instead.
+        Deprecated as of v0.25.0. Use ``plot_partial_effects_on_outcome`` instead.
         """
         warnings.warn("This method name is deprecated. Use `plot_partial_effects_on_outcome` instead.", DeprecationWarning)
         return plot_partial_effects_on_outcome(*args, **kwargs)