+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## Changelog
 
+#### 0.25.1 - 2020-08-01
+
+##### Bug fixes
+ - ok _actually_ ship the out-of-sample calibration code
+ - fix `labels=False` in `add_at_risk_counts`
+ - all for specific rows to be shown in `add_at_risk_counts`
+ - put `patsy` as a proper dependency.
+ - suppress some Pandas 1.1 warnings.
+
+
 #### 0.25.0 - 2020-07-27
 
 ##### New features